Sync encounter tables for route 3 and 4 with Pokemon Yellow
Since Sandshrew is being explicitly added to these encounter tables, I decided to just make them the same as in pokemon yellow - this makes pokemon on route 3 a slightly higher level, which may be good or bad, I'm not sure. I left the jigglypuff encounter data the same for route 3, which makes it a 10% encounter rate for jigglypuff, instead of an extra 5% for rattata and an extra 5% for spearow.