Swap safari zone tables for Pokemon Yellow version (For Scyther/Pinsir)
I was worrying about putting both scyther and pinsir in the safari zone in a satisfying way, but the solution is to just copy the encounter data from a game that has it all figured out. As a side effect, we can now capture Cubone and Marowak instead of Doduo and Dodrio, but I think that's probably fine.