Fix in-game trade evolutions, hopefully
There's a check for the first letter of a pokemon's name before it attempts to evolve them after an in-game trade. I changed haunter's letter from S (Spectre) to H, and added K for Kadabra and M for Machoke. I swapped out the Mr. Mime ingame trade because it also starts with the letter M, and it can already be caught elsewhere.