Change mew to level 7
This seemed like a good level, because the original mew glitch would generally let you encounter it at that level (this was because 7 was the default number for stat modifications - 1 would be at attack -6 and 13 would be at attack +6. that's a lot of growls or swords dances). And the second reason, now Mew can't use Transform and become a ditto when you catch it.