Shuffle trades around some more
Now that I no longer have to worry about the trade evo check, I can just have the mr mime trade again without worrying about the game trying to evolve it - apparently that was never a concern to begin with. So nidorina is out and mr mime is in. Furthermore, the machoke is now called ricky just like in pokemon yellow.